WIP: Simplify telemetry pausing / resume and fix blue bar bug
Previously, the notion of pause/resume of telemetry was actually
separate from enabling and disabling core location service updates.
Practically speaking however, there is no apparent reason to have two
sets of switches.

This is because pause only happens because a) the user turns off
telemetry via settings in the map or b) the host app only has
`when in use` permissions and the app is going to background. Otherwise,
(and assuming the host app has background location permissions from
iOS) we intend for telemetry to not pause and for locations services
to continue to run.

This commit simplifies the MGLMapboxEvents class's internal API
to help make it more clear that telemetry events (including location
services) will be paused or resumed, if required, when
NSUserDefaults are changed.

In addition, MGLMapboxEvents now observes application background and
foreground application lifecycle methods and pauses or unpauses
telemetry (and location services), if required. This allows us
to break the dependency that telemetry had on the map view class
in `wakeGL:`.

Finally, since MGLMapboxEvents is the only object that pauses or
unpauses telemetry, the public API it used to have for doing this
has been removed.